add function to return  results from pubmed
